requirements.txt
pdf2image
pillow
pytesseract
easyocr
numpy
torch
torchvision
